Fixing the UniversalGestureDetector
