+++ Package/lib/src/Widgets/Components/Communication/NumberPadComponent.dart
@@ -84,7 +84,7 @@ class _NumberPadComponentState extends State<NumberPadComponent> {
     return InteractiveSemanticsWrapper(
       properties: SemanticsWrapper.button(
           isEnabled: true,
-          label: number.toString,
+          label: '$number',
           hint: 'Inputs $number',
           isMutuallyExclusive: false),
       onInteract: () {
@@ -98,12 +98,6 @@ class _NumberPadComponentState extends State<NumberPadComponent> {
 
   @override
   Widget build(BuildContext context) {
-    List<Widget> digitButtons = [];
-
-    for (var element in digits) {
-      digitButtons.add(createNumberButton(element));
-    }
-
     var entryFieldBox = FloatingContainerElement(
       child: Container(
           decoration: InputBackingDecoration().buildBacking(),
