+++ Package/lib/src/Widgets/Elements/Accessibility/UniversalGestureDetectorElement.dart
@@ -1,6 +1,6 @@
 /// @nodoc
+import 'package:aureus/aureus.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 /// {@category Widgets}
 /// {@subCategory Elements}
@@ -11,7 +11,6 @@ import 'package:flutter/services.dart';
 /// that those with assistive tech can navigate
 
 class UniversalGestureDetector extends StatefulWidget {
-  /// The item that will have a 'floating glass' look behind it.
   final VoidCallback onDetect;
   final Widget child;
 
@@ -26,10 +25,52 @@ class UniversalGestureDetector extends StatefulWidget {
 }
 
 class _UniversalGestureDetectorState extends State<UniversalGestureDetector> {
+  Color borderColor = Colors.transparent;
+
+  void hideBorder() {
+    setState(() {
+      borderColor = Colors.transparent.withAlpha(0);
+    });
+  }
+
+  void showBorder() {
+    setState(() {
+      borderColor = coloration.accentColor();
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    return RawKeyboardListener(
-      focusNode: FocusNode(onKey: (node, event) {
+    return GestureDetector(
+      onTap: () {
+        widget.onDetect();
+      },
+      child: FocusableActionDetector(
+        onShowFocusHighlight: (isFocused) => {
+          isFocused ? showBorder() : hideBorder(),
+        },
+        actions: <Type, Action<Intent>>{
+          ActivateIntent: CallbackAction<Intent>(
+            onInvoke: (Intent intent) => widget.onDetect(),
+          ),
+        },
+        child: Container(
+          decoration: BoxDecoration(
+            border: Border.all(
+              style: BorderStyle.solid,
+              color: borderColor,
+            ),
+          ),
+          child: widget.child,
+        ),
+      ),
+    );
+  }
+}
+
+/*
+
+focusNode: FocusNode(onKey: (node, event) {
         if (event.isKeyPressed(LogicalKeyboardKey.enter)) {
           return KeyEventResult
               .handled; // prevent passing the event into the TextField
@@ -40,14 +81,10 @@ class _UniversalGestureDetectorState extends State<UniversalGestureDetector> {
         if (event.isKeyPressed(LogicalKeyboardKey.enter)) {
           // Do something
           widget.onDetect();
+        } else if (event.isKeyPressed(LogicalKeyboardKey.tab)) {
+          //shoud move focus to the next node
+          FocusScope.of(context).nextFocus();
         }
       },
-      child: GestureDetector(
-        onTap: () {
-          widget.onDetect();
-        },
-        child: widget.child,
-      ),
-    );
-  }
-}
+
+*/
